factor out DraggableList component
