+++ src/components/AliasForm.vue
@@ -12,24 +12,19 @@
         Include me in search results
       </label>
     </div>
-    <draggable
-      v-model="aliases"
-      :options="draggableOptions"
-      @start="startDrag"
-      @end="endDrag"
-    >
-      <transition-group :name="!draggingInProgress ? 'rearrange' : ''">
+    <DraggableList v-model="aliases" @reordered="reorder">
+      <template v-slot:listItems="eventListeners">
         <Alias
           v-for="(alias, index) in aliases"
           :key="alias.id"
           :alias="alias"
           :index="index"
           :disabled="alias.disabled"
-          @moved="aliasMoved"
+          @moved="eventListeners.moved"
           @destroy="destroy"
         />
-      </transition-group>
-    </draggable>
+      </template>
+    </DraggableList>
     <AddButton @create="create" aria-label="Add alias" />
   </form>
 </template>
@@ -37,35 +32,18 @@
 <script>
 import Alias from '@/components/Alias.vue';
 import AddButton from '@/components/AddButton.vue';
-import draggable from 'vuedraggable';
+import DraggableList from '@/components/DraggableList.vue';
 
 export default {
   name: 'AliasForm',
   components: {
     Alias,
     AddButton,
-    draggable,
-  },
-  props: {
-    draggableOptions: {
-      type: Object,
-      default() {
-        return {
-          animation: 200,
-          axis: 'y',
-          ghostClass: 'alias-placeholder',
-          handle: '.rearrange-handle',
-          items: '.alias',
-          scroll: true,
-          scrollSensitivity: 60,
-        };
-      },
-    },
+    DraggableList,
   },
   data() {
     return {
       aliases: [],
-      draggingInProgress: false,
       userInSearch: false,
     };
   },
@@ -80,76 +58,6 @@ export default {
     });
   },
   methods: {
-    /* Dragging-related methods */
-    startDrag() {
-      this.draggingInProgress = true;
-    },
-    endDrag() {
-      this.draggingInProgress = false;
-      this.reorder();
-    },
-    aliasMoved(e) {
-      /// e: {
-      ///   direction: 'up' | 'down',
-      ///   index: number,
-      ///   el: Element,
-      /// }
-      let newIndex;
-      const validUp = e.direction === 'up' && e.index > 0;
-      const validDown =
-        e.direction === 'down' && e.index < this.aliases.length - 1;
-      if (validUp) {
-        newIndex = e.index - 1;
-      } else if (validDown) {
-        newIndex = e.index + 1;
-      } else {
-        return;
-      }
-
-      const movingAlias = this.aliases[e.index];
-      this.aliases.splice(e.index, 1);
-      this.aliases.splice(newIndex, 0, movingAlias);
-
-      // For the duration of the transition as aliases are moved around,
-      // this code calls requestAnimationFrame and changes the window's scroll
-      // position per frame to make sure the moved alias stays in view.
-      let transitionEnded = false;
-      // this function does all the work and is added as a transitionstart
-      // event listener:
-      const keepElementInView = () => {
-        const belowView =
-          e.el.getBoundingClientRect().bottom > window.innerHeight;
-        const aboveView = e.el.getBoundingClientRect().top < 0;
-        if (belowView) {
-          e.el.scrollIntoView(false);
-        } else if (aboveView) {
-          e.el.scrollIntoView(true);
-        }
-        /* istanbul ignore else */
-        if (!transitionEnded) {
-          window.requestAnimationFrame(keepElementInView);
-        }
-      };
-
-      // This function removes the event listeners once they've outlived their
-      // usefulness and is added as a transitionend listener:
-      const cleanUpAfterTransition = () => {
-        transitionEnded = true;
-        e.el.removeEventListener('transitionstart', keepElementInView);
-        e.el.removeEventListener('transitionend', cleanUpAfterTransition);
-      };
-      // and this is where the listeners are actually added:
-      e.el.addEventListener('transitionstart', keepElementInView);
-      e.el.addEventListener('transitionend', cleanUpAfterTransition);
-
-      // This makes sure that the handle that was just used to move elements
-      // stays in focus:
-      this.$nextTick(() => {
-        e.handle.focus();
-      });
-
-      this.reorder();
-    },
     reorder() {
       return (
         this.reorderAliases()
@@ -158,11 +66,7 @@ export default {
       );
     },
     create() {
-      const newId =
-        Math.max.apply(
-          Math,
-          this.aliases.map(a => a.id),
-        ) + 1;
+      const newId = Math.max.apply(Math, this.aliases.map(a => a.id)) + 1;
       const newAlias = {
         id: newId,
         name: '',
@@ -232,8 +136,5 @@ export default {
 }
 .sortable-drag {
   opacity: 1;
-}
-.rearrange-move {
-  transition: transform 0.2s;
 }
 </style>
